CategoryService, deleteById method refactor

+++ store/src/main/java/com/ecommerce/store/service/CategoryService.java
@@ -1,16 +1,19 @@
 package com.ecommerce.store.service;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.ecommerce.store.domain.Category;
 import com.ecommerce.store.dto.CategoryDTO;
 import com.ecommerce.store.repository.CategoryRepository;
+import com.ecommerce.store.service.exception.DatabaseException;
 import com.ecommerce.store.service.exception.EntityNotFoundException;
 
 @Service
@@ -22,15 +25,13 @@ public class CategoryService {
 	@Transactional(readOnly = true)
 	public List<CategoryDTO> searchAll() {
 		List<Category> list = categoryRepository.findAll();
-		List<CategoryDTO> result = new ArrayList<>();
-		list.forEach(x -> result.add(new CategoryDTO(x)));
-		return result;
+		return list.stream().map(item -> new CategoryDTO(item)).collect(Collectors.toList());
 	}
 
 	@Transactional(readOnly = true)
 	public CategoryDTO findById(Long categoryId) {
 		Category result = categoryRepository.findById(categoryId)
-				.orElseThrow(() -> new EntityNotFoundException("Entity not found"));
+				.orElseThrow(() -> new EntityNotFoundException("Id not found " + categoryId));
 		return new CategoryDTO(result);
 	}
 
@@ -50,14 +51,16 @@ public class CategoryService {
 		return new CategoryDTO(entity);
 	}
 
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void deleteById(Long categoryId) {
-		
+		if(!categoryRepository.existsById(categoryId)) {
+			throw new EntityNotFoundException("Entity not found " + categoryId);
+		}
 		try {
 			categoryRepository.deleteById(categoryId);
 			
-		} catch (Error e) {
-			throw new EntityNotFoundException(e.getMessage());
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Integrity violation");
 		}
 		
 	}
